+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 * !!! Switch constants in ProcessQueueCommand from public to private
 * !!! Add interface to CrawlStrategy filename
+* !!! Switch getExtensionIcon() in TcaUtility from protected to private
 
 ### Fixed